pyproject.toml
[tool.poetry]
name = "All Logs"
version = "0.1.0"
description = ""
authors = ["Arthur Alves <arthuralves2p@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.30.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"

[tool.poetry.dependencies.fastapi]
extras = ["standard"]
version = "^0.115.8"

[tool.poetry.dependencies.tortoise-orm]
extras = ["psycopg"]
version = "^0.24.0"

[tool.poetry.dependencies.aerich]
extras = ["toml"]
version = "^0.8.1"

[tool.poetry.dependencies.uvicorn]
extras = ["standard"]
version = "^0.34.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"
mypy = "^1.15.0"

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["ALL"]
ignore = [
    "A002",
    "I",
    "D100",
    "D101",
    "D103",
    "D104",
    "D106",
    "D107",
    "D401",
    "S101",
]

[tool.ruff.per-file-ignores]
"app/tests/*" = ["ANN001", "ANN201"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.aerich]
tortoise_orm = "app.main.TORTOISE_ORM"
location = "./app/core/database/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
